+++ asus-nb-ctrl/src/lib.rs
@@ -8,10 +8,16 @@ pub mod ctrl_charge;
 ///
 pub mod ctrl_fan_cpu;
 ///
-pub mod ctrl_leds;
+pub mod ctrl_gfx;
 ///
+pub mod ctrl_leds;
+/// Control ASUS bios function such as boot sound, Optimus/Dedicated gfx mode
+pub mod ctrl_rog_bios;
 /// Laptop matching to determine capabilities
 pub mod laptops;
+
+/// Fetch all supported functions for the laptop
+pub mod supported;
 
 mod error;
 
@@ -37,4 +43,10 @@ pub trait CtrlTaskComplex {
     type A;
 
     fn do_task(&mut self, config: &mut Config, event: Self::A);
+}
+
+pub trait GetSupported {
+    type A;
+
+    fn get_supported() -> Self::A;
 }