+++ asus-nb/src/lib.rs
@@ -23,13 +23,15 @@ pub mod anime_dbus;
 /// Helper functions for the AniMe display
 pub mod anime_matrix;
 
+pub mod error;
+
 pub mod dbus_anime;
 pub mod dbus_charge;
 pub mod dbus_gfx;
 pub mod dbus_ledmode;
 pub mod dbus_profile;
-
-pub mod error;
+pub mod dbus_rogbios;
+pub mod dbus_supported;
 
 // static LED_INIT1: [u8; 2] = [0x5d, 0xb9];
 // static LED_INIT2: &str = "]ASUS Tech.Inc."; // ] == 0x5d